+++ libs/opengl/include/mrpt/opengl/TLightParameters.h
@@ -38,6 +38,22 @@ struct TLightParameters
 	float shadow_bias = 1e-5, shadow_bias_cam2frag = 1e-5,
 		  shadow_bias_normal = 1e-4;
 
+	/** Multiplier from eye distance to the length size of the squared area in
+	 * which to evaluate shadow casting by unidirectional light.
+	 * Unitless (meter/meter).
+	 * \note (New in MRPT 2.10.0)
+	 */
+	double eyeDistance2lightShadowExtension = 2.0;
+
+	/** Minimum extension (in [0,1] ratio of the light distance) of the shadow
+	 * map square ortho frustum. Should be roughly the maximum area of the
+	 * largest room for indoor environments to ensure no missing shadows in
+	 * distant areas.
+	 *
+	 * \note (New in MRPT 2.10.0)
+	 */
+	float minimum_shadow_map_extension_ratio = 0.03f;
+
 	void writeToStream(mrpt::serialization::CArchive& out) const;
 	void readFromStream(mrpt::serialization::CArchive& in);